+++ Roots-SDK-Routing-TestBed/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "23.0.1"
 
     defaultConfig {
-        applicationId "io.branch.appconnectordeeplinktestbed"
+        applicationId "io.branch.rootsroutingtestbed"
         minSdkVersion 14
         targetSdkVersion 23
         versionCode 1
