+++ Roots-SDK-Routing-TestBed/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">Roots-SDK-TestBed</string>
+    <string name="app_name">Roots-SDK-Routing-TestBed</string>
 
 
 </resources>
