+++ Roots-SDK-Routing-TestBed/src/io/branch/rootsroutingtestbed/DeepLinkTestActivity.java
@@ -1,4 +1,4 @@
-package io.branch.appconnectordeeplinktestbed;
+package io.branch.rootsroutingtestbed;
 
 import android.app.Activity;
 import android.os.Bundle;
