+++ Roots-SDK-TestBed/build.gradle
@@ -12,7 +12,7 @@ android {
     buildToolsVersion "23.0.1"
 
     defaultConfig {
-        applicationId "io.branch.appconnectortestbed"
+        applicationId "io.branch.rootstestbed"
         minSdkVersion 14
         targetSdkVersion 23
         versionCode 1
