+++ Roots-SDK/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="io.branch.appconnector"
+    package="io.branch.roots"
           android:versionCode="0"
           android:versionName="0.1">
