+++ Roots-SDK/src/io/branch/roots/AppLaunchConfig.java
@@ -1,4 +1,4 @@
-package io.branch.appconnector;
+package io.branch.roots;
 
 import android.net.Uri;
 import android.text.TextUtils;
