+++ Roots-SDK/src/io/branch/roots/Defines.java
@@ -1,4 +1,4 @@
-package io.branch.appconnector;
+package io.branch.roots;
 
 /**
  * Created by sojanpr on 4/12/16.
