+++ Roots-SDK/src/io/branch/roots/URLContent.java
@@ -1,4 +1,4 @@
-package io.branch.appconnector;
+package io.branch.roots;
 
 /**
  * Created by sojanpr on 4/28/16.
